requirements.txt
streamlit
transformers
beautifulsoup4
requests
python-dotenv
PyPDF2
python-docx
pdfplumber